backend/LM/requirements.txt
fastapi
pydantic
uvicorn
huggingface_hub
llama-cpp-python
python-dotenv